break out list-model.ts from metamodel.ts (using an agent)
